refactors out jobs TODO: refactor this out into jobs/installer TODO: This is a job/installer, rename it TODO: refactor this out into jobs/tracker
