+++ lib/granary.js
@@ -7,7 +7,7 @@ var conf = require('../config/config')().conf;
 // TODO: refactor the need to pass log to this, require it from the log file
 module.exports = function(log) {
     var Project = require('./project')(log);
-    var processor = require('../lib/job_processor')(log);
+    var installer = require('./jobs/installer')(log);
     log.debug('Redis Configuration', conf.get('redis'));
     // TODO: refactor this out into setup file
     var jobs = kue.createQueue({
@@ -23,9 +23,9 @@ module.exports = function(log) {
         return search = reds.createSearch(jobs.client.getKey('search'));
     }
 
-    processor.setup(jobs);
+    installer.setup(jobs);
 
-    var Tracker = require('../lib/tracker')(log, conf, jobs);
+    var Tracker = require('./jobs/tracker')(log, conf, jobs);
     var Auth = require('../lib/auth')(log, conf);
     var Routes = {};
