+++ lib/jobs/installer.js
@@ -5,20 +5,18 @@ var Promise = require('bluebird');
 var mkdirp = require('mkdirp');
 var Job = require('kue').Job;
 
-// TODO: refactor this out into jobs/installer
 // TODO: refactor the need to pass log to this, require it from the log file
 module.exports = function (log) {
 
   var bower = require('./installers/bower')(log);
   var npm = require('./installers/npm')(log);
 
-  // TODO: This is a job/installer, rename it
   // TODO: refactor this into not being a singleton
   function JobProcessor() {}
 
   JobProcessor.setup = function (jobs) {
 
-      var Project = require('./project')(log);
+      var Project = require('../project')(log);
 
     // TODO: one job at a time, npm can explode? (might need to investigate sandboxing job processing)
     // TODO: can Bower and NPM install the same time? (could be done if the process was threaded, might have an issue with concurrency)
