+++ lib/jobs/installers/bower.js
