+++ lib/jobs/installers/npm.js
