+++ lib/jobs/tracker.js
@@ -5,7 +5,6 @@ var Promise = require('bluebird');
 var rimraf = require('rimraf');
 var granary = require('granary')();
 
-// TODO: refactor this out into jobs/tracker
 // TODO: refactor the need to pass log to this, require it from the log file
 // TODO: refactor the need to pass conf to this, require it from the conf file
 module.exports = function (log, conf, jobs) {
@@ -38,7 +37,7 @@ module.exports = function (log, conf, jobs) {
         return granary.init(options)
           .then(
             function () {
-              log.debug('Tracker: Freight done.');
+              log.debug('Tracker: Granary done.');
               return new Promise(function (resolve, reject) {
                 rimraf(projectDirectory, function (err) {
                   if (err) {
